requirements.txt
# ChefConta - Sistema de Gestão Financeira
# Dependências do Projeto

# Interface Gráfica
customtkinter>=5.2.0
pillow>=10.0.0

# Banco de Dados
sqlalchemy>=2.0.0
alembic>=1.12.0

# Segurança e Autenticação
bcrypt>=4.0.0
pyjwt>=2.8.0

# Relatórios e Exportação
reportlab>=4.0.0
openpyxl>=3.1.0
pandas>=2.1.0

# Validação e Utilidades
python-dotenv>=1.0.0
python-dateutil>=2.8.0
validators>=0.22.0

# Gráficos e Visualização
matplotlib>=3.8.0

# Sistema
pyinstaller>=6.0.0
